requirements.txt
# Core dependencies
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# OpenAI API
openai>=1.0.0

# Environment variables
python-dotenv>=1.0.0

# Rate limiting
ratelimit>=2.2.1

# Retry logic
tenacity>=8.2.0

# URL parsing
validators>=0.22.0

# Logging
colorlog>=6.7.0

# Optional: Google Places API
googlemaps>=4.10.0
